Last change for OSX hopefully
Summary:
* getrlimit is only 64-bit on Linux
* surround the lib capabilities code in !defined(SKIP_USER_CHANGE), will remove
the need for it next week
* dlerror() is a const char* normally, fix warning
* cast malloc to char*
* include "base.h" for compatibility.h
* cast pthread_self() before doing modulus, its int64 on Mac
* Util::copy only works on Linux due to fdatasync

Test Plan:
make
make fast_tests

DiffCamp Revision: 142985
Reviewed By: hzhao
CC: hphp-diffs@lists, davidrecordon, hzhao
Revert Plan:
Ok

+++ src/util/light_process.cpp
@@ -418,7 +418,7 @@ void LightProcess::runShadow(int fdin, int fdout) {
 }
 
 int LightProcess::GetId() {
-  return (int)pthread_self() % g_procs.size();
+  return (long)pthread_self() % g_procs.size();
 }
 
 FILE *LightProcess::popen(const char *cmd, const char *type,
